+++ IMIS_DataEntity/EntityClass/AccLastYrPeski.cs
@@ -5,11 +5,11 @@ namespace IMIS_DataEntity.EntityClass
 {
     public partial class AccLastYrPeski
     {
-        public decimal Sn { get; set; }
-        public decimal? PersonTypeId { get; set; }
-        public decimal? PersonId { get; set; }
+        public int Sn { get; set; }
+        public int? PersonTypeId { get; set; }
+        public int? PersonId { get; set; }
         public string AdvanceType { get; set; }
-        public decimal? SubModuleId { get; set; }
-        public double? TotalPesRem { get; set; }
+        public int? SubModuleId { get; set; }
+        public float? TotalPesRem { get; set; }
     }
 }
