+++ IMIS_DataEntity/EntityClass/AccPaymentCredit.cs
@@ -5,12 +5,12 @@ namespace IMIS_DataEntity.EntityClass
 {
     public partial class AccPaymentCredit
     {
-        public decimal Id { get; set; }
-        public decimal? PaymentId { get; set; }
+        public int Id { get; set; }
+        public int? PaymentId { get; set; }
         public string IsOfLastYear { get; set; }
-        public decimal? AccId { get; set; }
-        public decimal? BudgetAccId { get; set; }
-        public decimal? BudgetSourceId { get; set; }
-        public double? Amount { get; set; }
+        public int? AccId { get; set; }
+        public int? BudgetAccId { get; set; }
+        public int? BudgetSourceId { get; set; }
+        public float? Amount { get; set; }
     }
 }
