+++ IMIS_DataEntity/EntityClass/AccPaymentDeduction.cs
@@ -5,14 +5,14 @@ namespace IMIS_DataEntity.EntityClass
 {
     public partial class AccPaymentDeduction
     {
-        public decimal Id { get; set; }
-        public decimal? PaymentId { get; set; }
-        public decimal? AccId { get; set; }
-        public decimal? BudgetAccId { get; set; }
-        public double? Amount { get; set; }
+        public int Id { get; set; }
+        public int? PaymentId { get; set; }
+        public int? AccId { get; set; }
+        public int? BudgetAccId { get; set; }
+        public float? Amount { get; set; }
         public string Status { get; set; }
         public string Deductiontype { get; set; }
-        public decimal? BudgetSourceId { get; set; }
+        public int? BudgetSourceId { get; set; }
 
         public virtual AccAccMaster Acc { get; set; }
         public virtual AccAccMaster BudgetAcc { get; set; }
