+++ IMIS_DataEntity/EntityClass/AccReceivable.cs
@@ -5,16 +5,16 @@ namespace IMIS_DataEntity.EntityClass
 {
     public partial class AccReceivable
     {
-        public decimal? Sn { get; set; }
-        public decimal? Taxpayerid { get; set; }
+        public int? Sn { get; set; }
+        public int? Taxpayerid { get; set; }
         public string Fiscalyear { get; set; }
-        public double? Nettaxamount { get; set; }
-        public double? Fineamount { get; set; }
-        public double? Additionalcharges { get; set; }
-        public double? Reductionalcharges { get; set; }
+        public float? Nettaxamount { get; set; }
+        public float? Fineamount { get; set; }
+        public float? Additionalcharges { get; set; }
+        public float? Reductionalcharges { get; set; }
         public string Taxtype { get; set; }
         public int? Type { get; set; }
-        public decimal? Vchrid { get; set; }
+        public int? Vchrid { get; set; }
         public string VchrDate { get; set; }
     }
 }
