+++ IMIS_DataEntity/EntityClass/AccUnauthorised.cs
@@ -5,15 +5,15 @@ namespace IMIS_DataEntity.EntityClass
 {
     public partial class AccUnauthorised
     {
-        public decimal UnauthorisedId { get; set; }
-        public decimal? SubModuleId { get; set; }
-        public decimal? WorkId { get; set; }
-        public decimal? UnauthorisedDafaId { get; set; }
-        public decimal? UnauthorisedAmt { get; set; }
+        public int UnauthorisedId { get; set; }
+        public int? SubModuleId { get; set; }
+        public int? WorkId { get; set; }
+        public int? UnauthorisedDafaId { get; set; }
+        public int? UnauthorisedAmt { get; set; }
         public string InspectDate { get; set; }
         public string InspectDetails { get; set; }
-        public decimal? InspectAmt { get; set; }
-        public decimal? FilledbyEmpid { get; set; }
-        public decimal? VerifiedbyEmpid { get; set; }
+        public int? InspectAmt { get; set; }
+        public int? FilledbyEmpid { get; set; }
+        public int? VerifiedbyEmpid { get; set; }
     }
 }
