+++ IMIS_DataEntity/EntityClass/AccUserComiteMstr.cs
@@ -5,10 +5,10 @@ namespace IMIS_DataEntity.EntityClass
 {
     public partial class AccUserComiteMstr
     {
-        public decimal Id { get; set; }
+        public int Id { get; set; }
         public string NepName { get; set; }
         public string EngName { get; set; }
-        public decimal? Ward { get; set; }
+        public int? Ward { get; set; }
         public string Address { get; set; }
         public string Email { get; set; }
         public string Phones { get; set; }
