+++ IMIS_DataEntity/EntityClass/AccWorkFundsDetail.cs
@@ -5,13 +5,13 @@ namespace IMIS_DataEntity.EntityClass
 {
     public partial class AccWorkFundsDetail
     {
-        public decimal FundDetailId { get; set; }
-        public decimal? WorkId { get; set; }
-        public decimal? ProviderId { get; set; }
-        public double? FundedAmt { get; set; }
-        public double? FundedPercent { get; set; }
-        public decimal? Isjanasaramdha { get; set; }
-        public decimal? Isownoffice { get; set; }
+        public int FundDetailId { get; set; }
+        public int? WorkId { get; set; }
+        public int? ProviderId { get; set; }
+        public float? FundedAmt { get; set; }
+        public float? FundedPercent { get; set; }
+        public int? Isjanasaramdha { get; set; }
+        public int? Isownoffice { get; set; }
 
         public virtual AccFundProvidrMstr Provider { get; set; }
         public virtual AccWorkMaster Work { get; set; }
