+++ IMIS_DataEntity/EntityClass/Country.cs
@@ -5,7 +5,7 @@ namespace IMIS_DataEntity.EntityClass
 {
     public partial class Country
     {
-        public decimal? Id { get; set; }
+        public int? Id { get; set; }
         public string Nepname { get; set; }
         public string Engname { get; set; }
         public string Code { get; set; }
