+++ IMIS_DataEntity/EntityClass/Ddc.cs
@@ -5,11 +5,11 @@ namespace IMIS_DataEntity.EntityClass
 {
     public partial class Ddc
     {
-        public decimal? Id { get; set; }
+        public int? Id { get; set; }
         public string Nepname { get; set; }
         public string Engname { get; set; }
         public string Code { get; set; }
         public string Remarks { get; set; }
-        public decimal? Districtid { get; set; }
+        public int? Districtid { get; set; }
     }
 }
