+++ IMIS_DataEntity/EntityClass/Fywiserentdarrate.cs
@@ -5,11 +5,11 @@ namespace IMIS_DataEntity.EntityClass
 {
     public partial class Fywiserentdarrate
     {
-        public decimal Id { get; set; }
-        public decimal? Rateid { get; set; }
+        public int Id { get; set; }
+        public int? Rateid { get; set; }
         public string Ratename { get; set; }
-        public double? RentPer { get; set; }
-        public decimal? RentTypeid { get; set; }
+        public float? RentPer { get; set; }
+        public int? RentTypeid { get; set; }
         public string FiscalYear { get; set; }
     }
 }
