+++ IMIS_DataEntity/EntityClass/InvApplicantRate.cs
@@ -5,9 +5,9 @@ namespace IMIS_DataEntity.EntityClass
 {
     public partial class InvApplicantRate
     {
-        public decimal Id { get; set; }
-        public decimal? Applicantid { get; set; }
-        public decimal? ItemDetailId { get; set; }
-        public double? Amount { get; set; }
+        public int Id { get; set; }
+        public int? Applicantid { get; set; }
+        public int? ItemDetailId { get; set; }
+        public float? Amount { get; set; }
     }
 }
