+++ IMIS_DataEntity/EntityClass/InvDept.cs
@@ -10,10 +10,10 @@ namespace IMIS_DataEntity.EntityClass
             InvRoomMst = new HashSet<InvRoomMst>();
         }
 
-        public decimal DeptId { get; set; }
+        public int DeptId { get; set; }
         public string NameNp { get; set; }
         public string NameEn { get; set; }
-        public decimal? OfficeId { get; set; }
+        public int? OfficeId { get; set; }
         public string IsActive { get; set; }
 
         public virtual ICollection<InvRoomMst> InvRoomMst { get; set; }
