+++ IMIS_DataEntity/EntityClass/InvGoodsRecMast.cs
@@ -11,21 +11,21 @@ namespace IMIS_DataEntity.EntityClass
             InvItemDetails = new HashSet<InvItemDetails>();
         }
 
-        public decimal Id { get; set; }
+        public int Id { get; set; }
         public string DakhilaId { get; set; }
-        public decimal? PurMasterId { get; set; }
-        public decimal? RepairId { get; set; }
+        public int? PurMasterId { get; set; }
+        public int? RepairId { get; set; }
         public string DakhilaDt { get; set; }
-        public decimal? WarehouseId { get; set; }
+        public int? WarehouseId { get; set; }
         public string InvBillNo { get; set; }
         public string InvDate { get; set; }
         public string DetailsNep { get; set; }
         public string DetailsEng { get; set; }
-        public decimal? PrepBy { get; set; }
+        public int? PrepBy { get; set; }
         public string PrepDt { get; set; }
-        public decimal? CheckBy { get; set; }
+        public int? CheckBy { get; set; }
         public string CheckDt { get; set; }
-        public decimal? AcceptBy { get; set; }
+        public int? AcceptBy { get; set; }
         public string AcceptDt { get; set; }
         public DateTime? DakhilaDtEng { get; set; }
         public DateTime? InvDateEng { get; set; }
@@ -33,7 +33,7 @@ namespace IMIS_DataEntity.EntityClass
         public DateTime? CheckDtEng { get; set; }
         public DateTime? AcceptDtEng { get; set; }
         public string FiscalYear { get; set; }
-        public decimal? PaymentId { get; set; }
+        public int? PaymentId { get; set; }
 
         public virtual PisEmployeeMaster AcceptByNavigation { get; set; }
         public virtual PisEmployeeMaster CheckByNavigation { get; set; }
